1,Add Utclient; 2,Server client bug fix

+++ h/server/Server.h
@@ -25,7 +25,7 @@ struct Server {
 
 typedef struct ServerParam {
         RequestHandler  *request_handler;
-        uint8_t         request_handler_length;
+        uint16_t        request_handler_length;
         ThreadPool      *worker_tp;
         int             port;
         size_t          read_buffer_size;
